requirements.txt
aiogram==2.25.2
aiohttp==3.10.11
aiosignal==1.3.1
python-dotenv==0.21.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0